BackEnd/app/config.py
import os
from dotenv import load_dotenv  # 👈 add this

load_dotenv()

class Config:
    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URL")
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SECRET_KEY = os.getenv("SECRET_KEY", "default_secret")
    UPLOAD_FOLDER = 'uploads'


