BackEnd/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.3
python-dateutil

pymongo==4.4.0
cassandra-driver==3.29.3


